Support React Router v5 in RouterToUrlQuery
React Router v5 switched from the old style React context (this.context)
to the new React context style (<Context.Provider>, <Context.Consumer>).
This change caused RouterToUrlQuery to stop working with React Router.

I've upgraded RouterToUrlQuery to use the provided ReactRouter context
object to supply the router instance to RouterToUrlQuery. One
requirement of this changed was to add react-router to devDependencies
and peerDependencies. I'm not sure NPM has the right dependency type for
this, peerDependencies seemed like the least intrusive.

I've also updated the Webpack config to mark React, ReactDOM and React
Router as external dependencies. This is a requirement to ensure this
package does not ship it's own version of React Router; this would cause
two contexts to be created and the package will not be able to retrieve
the router from context. I honestly have no idea what effect this might
have on the various builds generated. All I know is that it works for us
in our Webpack setup.